% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_preliminary_linear_range.R
\name{calcCV}
\alias{calcCV}
\title{PLR: calculate CV for one concentration level}
\usage{
calcCV(x)
}
\arguments{
\item{x}{\strong{data.frame} \cr Data.frame containing data for a specific concentration level.}
}
\value{
\strong{numeric(1)} \cr Coefficient of variation (CV) for the given concentration level.
}
\description{
PLR: calculate CV for one concentration level
}
