% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{calcPerBias}
\alias{calcPerBias}
\title{FLR: calculate percent bias for a specific concentration level}
\usage{
calcPerBias(x, LMfit, expConc)
}
\arguments{
\item{x}{data frame for a specific concentration level}

\item{LMfit}{linear model fit}

\item{expConc}{expected (known) concentration level}
}
\description{
FLR: calculate percent bias for a specific concentration level
}
