% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_response_factors.R
\name{calcResponseFactors}
\alias{calcResponseFactors}
\title{Calculate Response factors}
\usage{
calcResponseFactors(x, intercept, expConc)
}
\arguments{
\item{intercept}{Intercept of the linear model}

\item{expConc}{}
}
\description{
Function, which calculates a response factor for a single data point
}
\details{
Formula obtained from:  Green, J. M., A practical guide to analytical method validation.
}
