% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_response_factors.R
\name{calcRFMeans}
\alias{calcRFMeans}
\title{Calculate Response factors}
\usage{
calcRFMeans(x)
}
\arguments{
\item{x}{\strong{list} \cr Result of \code{\link{calcRFLevels}} \cr List of response factor values for each concentration level.}
}
\value{
vector of mean response factor values for the different concentration levels.
}
\description{
The function returns mean response factor values
}
\examples{
#' data(D_MFAP4)
D_MFAP4_cleaned <- cleanData(D_MFAP4, minReplicates = 3)
RES_PLR <- calculate_PLR(D_MFAP4_cleaned,
    cv_thres = 10,
    calcContinuousPrelimRanges = TRUE
)
RES_FLR <- calculate_FLR(RES_PLR$dataPrelim)

RES_RF <- calcRFLevels(D_MFAP4_cleaned, mod = RES_FLR$mod)
calcRFMeans(RES_RF)
}
