% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{checkFinalRange}
\alias{checkFinalRange}
\title{FLR: checks if final linear range has been reached (compare average percent bias with threshold)}
\usage{
checkFinalRange(perBiasInfo, perBiasThres = 10)
}
\arguments{
\item{perBiasInfo}{result of calcPerBiasAvgSDCV()}

\item{perBiasThres}{threshold for average percent bias}
}
\value{
logical(1) TRUE if bost lowest and highest concentration level passed the check
(and the final linear range is reached), FALSE otherwise
}
\description{
FLR: checks if final linear range has been reached (compare average percent bias with threshold)
}
