% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D_ALB}
\alias{D_ALB}
\title{Targeted Proteomics data set Albumin}
\format{
\subsection{\code{D_ALB}}{

A data frame with 21 rows and 2 columns:
\describe{
\item{Concentration}{Concentration}
\item{Measurement}{Measurement}
}
}
}
\source{
TODO
}
\usage{
D_ALB
}
\description{
Details TODO
}
\keyword{datasets}
