% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_preliminary_linear_range.R
\name{.calcContPrelimRanges}
\alias{.calcContPrelimRanges}
\title{PLR: Calculate key features for the existing continuous preliminary ranges}
\usage{
.calcContPrelimRanges(index)
}
\arguments{
\item{index}{\strong{integer} \cr Integer vector containing indices of
concentration levels with CV < threshold.}
}
\value{
Data.frame with one range in each row. Contains the columns
"startPoints", "endPoints", and "extent".
}
\description{
PLR: Calculate key features for the existing continuous preliminary ranges
}
