% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{.calcPerBiasLevels}
\alias{.calcPerBiasLevels}
\title{FLR: calculate list of percent bias values for all concentration levels}
\usage{
.calcPerBiasLevels(x, LMfit)
}
\arguments{
\item{x}{\strong{list of data.frames} \cr List of data frames (one entry for each
concentration level), e.g. output "dataPrelim" from
\code{\link{calculate_PLR}}.}

\item{LMfit}{\strong{lm object} \cr Linear model fit as calculated by
\code{\link{.calcLinearModel}}.}
}
\value{
List with percent bias values for each data point per
concentration level
}
\description{
FLR: calculate list of percent bias values for all concentration levels
}
