% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{.calcWeights}
\alias{.calcWeights}
\title{FLR: calculate weights for linear model for a specific concentration level}
\usage{
.calcWeights(x, weightingMethod = "1/x^2")
}
\arguments{
\item{x}{\strong{data.frame} \cr Data.frame containing data for a specific
concentration level.}

\item{weightingMethod}{\strong{character(1)} \cr Method for weighting (currently
"1/x", "1/x^2" and "None" are supported, default is "1/x^2").}
}
\value{
Vector of a constant weights for each measurement in this
concentration level.
}
\description{
FLR: calculate weights for linear model for a specific concentration level
}
