% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_preprocessing.R
\name{.checkNumberReplicates}
\alias{.checkNumberReplicates}
\title{Data preprocessing: Helper function to check for sufficient number of
replicates for a specific concentration level}
\usage{
.checkNumberReplicates(x, data, minReplicates)
}
\arguments{
\item{x}{\strong{numeric(1)} \cr concentration level to check}

\item{data}{\strong{list of data.frames} \cr list of data.frames (each dataframe
contains data for a specific concentration level)}

\item{minReplicates}{\strong{integer(1)} \cr minimal number of data points per
concentration level}
}
\value{
\strong{logical(1)} \cr TRUE if there are enough replicates, else FALSE
}
\description{
Data preprocessing: Helper function to check for sufficient number of
replicates for a specific concentration level
}
