% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_plotting.R
\name{.prepareAnnotationData}
\alias{.prepareAnnotationData}
\title{Prepare annotation data}
\usage{
.prepareAnnotationData(D_calib)
}
\arguments{
\item{D_calib}{\strong{data.frame} \cr Result of \code{\link{.prepareCalibData}}}
}
\value{
Dataframe with annotation data for the plot.
}
\description{
Prepare annotation data
}
