% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_plotting.R
\name{.prepareCalibData}
\alias{.prepareCalibData}
\title{Prepare calibration data}
\usage{
.prepareCalibData(CC_RES)
}
\arguments{
\item{CC_RES}{\strong{list} Result object of \code{\link{CalibraCurve}}.}
}
\value{
Dataframe with calibration data for plotting.
}
\description{
Prepare calibration data
}
