% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_plotting.R
\name{.prepareCurveData}
\alias{.prepareCurveData}
\title{Prepare curve data}
\usage{
.prepareCurveData(CC_RES, D_calib)
}
\arguments{
\item{CC_RES}{\strong{list} \cr Results of \code{\link{CalibraCurve}}. Each list
element is the result of \code{\link{calc_single_curve}}}

\item{D_calib}{\strong{data.frame} \cr Result of \code{\link{.prepareCalibData}}}
}
\value{
Dataframe with data points on a grid to plot the calibration curve.
}
\description{
Prepare curve data
}
