% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_writing_results.R
\name{.saveCCResult}
\alias{.saveCCResult}
\title{Save results of CalibraCurve}
\usage{
.saveCCResult(CC_res, output_path, suffix = "")
}
\arguments{
\item{CC_res}{\strong{list} Result object of \code{\link{CalibraCurve}}.}

\item{output_path}{\strong{character(1)} \cr Path to the output directory.}

\item{suffix}{\strong{character(1)} \cr Suffix for the output files, ideally
starting with "_" (default is "").}
}
\value{
Returns nothing, but the function saves the results to the specified
output path.
}
\description{
Save results of CalibraCurve
}
