% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{.selctConcLevel}
\alias{.selctConcLevel}
\title{FLR: selects the highest or lowest concentration level for removal}
\usage{
.selctConcLevel(x, perBiasT = 20, consPerBiasCV = TRUE, perBiasDistT = 10)
}
\arguments{
\item{x}{\strong{data.frame} Result of \code{\link{.calcPerBiasAvgSDCV}}.}

\item{perBiasT}{\strong{numeric(1)} \cr Threshold for average percent bias in
percent, default is 20.}

\item{consPerBiasCV}{consider CV? default is TRUE}

\item{perBiasDistT}{\strong{numeric(1)} \cr Threshold for the difference in
average percent bias in percent (for lower differences, CV will be
considered), default is 10. Only used if consPerBiasCV is TRUE.}
}
\value{
TRUE, if lowest concentration will be removed, FALSE if highest
will be removed
}
\description{
FLR: selects the highest or lowest concentration level for removal
}
