% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_preprocessing.R
\name{filterConcentrationLevel}
\alias{filterConcentrationLevel}
\title{Data preprocessing: Helper function to select all rows from a specific concentration level}
\usage{
filterConcentrationLevel(x, rawData)
}
\arguments{
\item{x}{\strong{numeric(1)} \cr concentration level to select}

\item{rawData}{\strong{data.frame} \cr data set to be filtered (e.g. result of \code{\link{readData}})}
}
\value{
data.frame
}
\description{
Data preprocessing: Helper function to select all rows from a specific concentration level
}
