% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RES_ALB}
\alias{RES_ALB}
\title{CalibraCurve results for data set Albumin}
\format{
\subsection{\code{RES_ALB}}{

Results from CalibraCurve calculations.
}
}
\usage{
RES_ALB
}
\description{
CalibraCurve results for the data set \code{\link{D_ALB}}. Default parameters for CalibraCurve were used.
}
\keyword{datasets}
