% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RES_Apolipoprotein}
\alias{RES_Apolipoprotein}
\title{CalibraCurve results for data set Apolipoprotein}
\format{
\subsection{\code{RES_Apolipoprotein}}{

Results from CalibraCurve calculations.
}
}
\usage{
RES_Apolipoprotein
}
\description{
CalibraCurve results for the data set \code{\link{D_Apolipoprotein}}. Default parameters for CalibraCurve were used.
}
\keyword{datasets}
