% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RES_MFAP4}
\alias{RES_MFAP4}
\title{CalibraCurve results for data set MFAP4}
\format{
\subsection{\code{RES_MFAP4}}{

Results from CalibraCurve calculations.
}
}
\usage{
RES_MFAP4
}
\description{
CalibraCurve results for the data set \code{\link{D_MFAP4}}. Default parameters for CalibraCurve were used.
}
\keyword{datasets}
