% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_preprocessing.R
\name{selConcentrationLevels}
\alias{selConcentrationLevels}
\title{Data preprocessing: Select all rows with identical concentrations}
\usage{
selConcentrationLevels(x, rawData)
}
\arguments{
\item{x}{concentration level to check}

\item{rawData}{data set to be filtered}
}
\description{
Data preprocessing: Select all rows with identical concentrations
}
