% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{selctConcLevel}
\alias{selctConcLevel}
\title{FLR: selects a concentration level for subsequent removal}
\usage{
selctConcLevel(x, consPerBiasCV, perBiasT, perBiasDistT)
}
\arguments{
\item{x}{result of calcPerBiasAvgSDCV()}

\item{consPerBiasCV}{consider CV?}

\item{perBiasT}{threshold for average percent bias}

\item{perBiasDistT}{threshold for the difference in average percent bias (for lower differences, CV will be considered)}
}
\value{
logical(1): TRUE, if lowest concentration will be removed, FALSE if highest will be removed
}
\description{
FLR: selects a concentration level for subsequent removal
}
