% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_writing_results.R
\name{assemble_results}
\alias{assemble_results}
\title{Title}
\usage{
assemble_results(
  X,
  dataValidated,
  cv_thres = 20,
  PLR_res,
  resFacDataV,
  avgResFacDataV,
  FLR_res,
  mod,
  RfThresL = 80,
  RfThresU = 120,
  substance = "substance1"
)
}
\arguments{
\item{mod}{}
}
\description{
Title
}
