% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEW_00_main_CalibraCurve.R
\name{calc_single_curve}
\alias{calc_single_curve}
\title{Title}
\usage{
calc_single_curve(
  single_path,
  output_path = NULL,
  conc_col,
  meas_col,
  substance = "substance1",
  suffix = paste0("_", substance),
  filetype = "xlsx",
  sep = ",",
  dec = ".",
  header = TRUE,
  na.strings = c("NA", "NaN", "Filtered", "#NV"),
  sheet = 1,
  min_replicates = 3,
  cv_thres = 20,
  calcContinuousPrelimRanges = TRUE,
  weightingMethod = "1/x^2",
  centralTendencyMeasure = "mean",
  perBiasThres = 30,
  considerPerBiasCV = TRUE,
  perBiasDistThres = 10,
  RfThresL = 80,
  RfThresU = 120
)
}
\arguments{
\item{RfThresU}{}
}
\description{
Title
}
