% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_preliminary_linear_range.R
\name{calcContPrelimRanges}
\alias{calcContPrelimRanges}
\title{PLR: Calculate key features for the existing continuous preliminary ranges}
\usage{
calcContPrelimRanges(x)
}
\arguments{
\item{x}{data frame}
}
\description{
PLR: Calculate key features for the existing continuous preliminary ranges
}
