% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_preliminary_linear_range.R
\name{calcCV}
\alias{calcCV}
\title{PLR: calculate CV for each concentration level}
\usage{
calcCV(x)
}
\arguments{
\item{x}{data frame}
}
\description{
PLR: calculate CV for each concentration level
}
