% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{calcLinearModel}
\alias{calcLinearModel}
\title{FLR: calculate weighted or unweighted linear model}
\usage{
calcLinearModel(x, weights = NULL)
}
\arguments{
\item{x}{list of data frames (one entry for each concentration level)}

\item{weights}{vector of weights (default is NULL and will result in an unweighted model)}
}
\description{
FLR: calculate weighted or unweighted linear model
}
