% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{calcPerBias}
\alias{calcPerBias}
\title{FLR: calculate percent bias for a specific concentration level}
\usage{
calcPerBias(x, LMfit, expConc)
}
\arguments{
\item{x}{\strong{data.frame} \cr Data.frame containing data for a specific concentration level.}

\item{LMfit}{\strong{lm object} \cr Linear model fit as calculated by \code{\link{calcLinearModel}}.}

\item{expConc}{\strong{numeric(1)} \cr Expected (known) concentration level.}
}
\value{
Vector of percent bias values for each data point in this concentration level.
}
\description{
FLR: calculate percent bias for a specific concentration level
}
