% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{calcPerBiasAvgSDCV}
\alias{calcPerBiasAvgSDCV}
\title{FLR: calculate average, SD and CV percent bias for each concentration level}
\usage{
calcPerBiasAvgSDCV(x, method = "mean")
}
\arguments{
\item{x}{result of calcPerBiasLevels()}

\item{method}{method: "mean" or "median", default is "mean"}
}
\value{
data frame with 3 columns: avgPerBias, stdDevPerBias, CV_PerBias
each row is one concentration level
}
\description{
FLR: calculate average, SD and CV percent bias for each concentration level
}
