% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{calcPerBiasLevels}
\alias{calcPerBiasLevels}
\title{FLR: calculate list of percent bias values for all concentration levels}
\usage{
calcPerBiasLevels(x, LMfit)
}
\arguments{
\item{x}{list of data frames (one list entry for each concentration level)}

\item{LMfit}{linear model fit}
}
\value{
list of vectors with percent bias values for each data point per concentration level
}
\description{
FLR: calculate list of percent bias values for all concentration levels
}
