% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_response_factors.R
\name{calcRFLevels}
\alias{calcRFLevels}
\title{Calculate Response factors}
\usage{
calcRFLevels(x, mod)
}
\arguments{
\item{mod}{linear model object}
}
\description{
Final linear range: Function, which returns a list with response factor values for a data set (given as list)
}
