% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_response_factors.R
\name{calcRFMeans}
\alias{calcRFMeans}
\title{Calculate Response factors}
\usage{
calcRFMeans(x)
}
\arguments{
\item{x}{\strong{list} \cr Result of \code{\link{calcRFLevels}} \cr List of response factor values for each concentration level.}
}
\value{
vector of mean response factor values for the different concentration levels.
}
\description{
The function returns mean response factor values
}
