% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_preliminary_linear_range.R
\name{calculate_PLR}
\alias{calculate_PLR}
\title{Calulate preliminary linear range (PLR)}
\usage{
calculate_PLR(dataValidated, cv_thres = 20, calcContinuousPrelimRanges = TRUE)
}
\arguments{
\item{dataValidated}{data frame cleaned by cleanData()}

\item{cv_thres}{threshold for CV per concentration level (default is 20)}

\item{calcContinuousPrelimRanges}{}
}
\value{
list
}
\description{
Calulate preliminary linear range (PLR)
}
