% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{calcWeights}
\alias{calcWeights}
\title{FLR: calculate weights for linear model for a specific concentration level}
\usage{
calcWeights(x, weightMet = "1/x^2")
}
\arguments{
\item{x}{data frame}

\item{weightMet}{method for weighting (currently 1/x and 1/x^2 are supported, default is 1/x^2)}
}
\description{
FLR: calculate weights for linear model for a specific concentration level
}
