% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_main_CalibraCurve.R
\name{CalibraCurve}
\alias{CalibraCurve}
\title{CalibraCurve}
\usage{
CalibraCurve(
  path,
  conc_col,
  meas_col,
  filetype = "xlsx",
  sep = ",",
  dec = ";",
  header = TRUE,
  na.strings = c("NA", "NaN", "Filtered", "#NV"),
  sheet = 1,
  min_replicates = 3,
  cv_thres = 20,
  calcContinuousPrelimRanges = TRUE,
  weightingMethod = "1/x^2",
  centralTendencyMeasure = "mean",
  perBiasThres = 20,
  considerPerBiasCV = TRUE,
  perBiasDistThres = 10,
  RfThresL = 80,
  RfThresU = 120,
  substance = "substance1"
)
}
\arguments{
\item{perBiasDistThres}{}
}
\description{
CalibraCurve
}
