% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_main_CalibraCurve.R
\name{CalibraCurve}
\alias{CalibraCurve}
\title{CalibraCurve}
\usage{
CalibraCurve(
  data_path,
  output_path = NULL,
  conc_col,
  meas_col,
  filetype = "xlsx",
  sep = ",",
  dec = ";",
  header = TRUE,
  na.strings = c("NA", "NaN", "Filtered", "#NV"),
  sheet = 1,
  min_replicates = 3,
  cv_thres = 20,
  calcContinuousPrelimRanges = TRUE,
  weightingMethod = "1/x^2",
  centralTendencyMeasure = "mean",
  perBiasThres = 20,
  considerPerBiasCV = TRUE,
  perBiasDistThres = 10,
  RfThresL = 80,
  RfThresU = 120,
  substance = "substance1"
)
}
\arguments{
\item{data_path}{\strong{character(1)} \cr Path to the data file (.csv, .txt or .xlsx file).}

\item{output_path}{\strong{character(1)} \cr Folder to save results (table and plots).}

\item{conc_col}{\strong{integer(1)} \cr Column number of the concentration values.}

\item{meas_col}{\strong{integer(1)} \cr Column number of the concentration values.}

\item{filetype}{\strong{character(1)} \cr Type of input file: "csv" or "txt" or "xlsx".}

\item{sep}{\strong{character(1)} \cr The field separator, e.g. " " for blanks, "," for comma or "\t" for tab.}

\item{dec}{\strong{character(1)} \cr Decimal separator, e.g. "," for comma or "." for dot.}

\item{header}{\strong{logical(1)} \cr If TRUE, first line is counted as column names.}

\item{na.strings}{\strong{character} \cr Character vector of strings which are to be interpreted as NA.}

\item{sheet}{\strong{integer(1)} \cr Sheet number (only needed for xlsx files, default is to use the first sheet).}

\item{min_replicates}{\strong{integer(1)} \cr Minimal number of replicates/data points per concentration level.
Concentration levels with too few data points will be removed.}

\item{substance}{}
}
\description{
CalibraCurve
}
