% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_final_linear_range.R
\name{checkFinalRange}
\alias{checkFinalRange}
\title{FLR: checks if final linear range has been reached (compare average percent bias with threshold)}
\usage{
checkFinalRange(perBiasInfo, perBiasThres = 20)
}
\arguments{
\item{perBiasInfo}{\strong{data.frame} Result of \code{\link{calcPerBiasAvgSDCV}}.}

\item{perBiasThres}{\strong{numeric(1)} \strong{numeric(1)} \cr Threshold for average percent bias in percent, default is 20.}
}
\value{
TRUE if both lowest and highest concentration level passed the check
(and the final linear range is reached), FALSE otherwise
}
\description{
FLR: checks if final linear range has been reached (compare average percent bias with threshold)
}
