% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_preprocessing.R
\name{checkNumberReplicates}
\alias{checkNumberReplicates}
\title{Data preprocessing: Check for sufficient number of replicates}
\usage{
checkNumberReplicates(x, data, minNumber)
}
\arguments{
\item{x}{concentration level to check}

\item{data}{data set to be filtered}

\item{minNumber}{minimal number of data points per concentration level}
}
\description{
Data preprocessing: Check for sufficient number of replicates
}
