% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_preprocessing.R
\name{cleanData}
\alias{cleanData}
\title{Clean data}
\usage{
cleanData(rawData, min_replicates)
}
\arguments{
\item{rawData}{data set to be cleaned, result of readData.}

\item{min_replicates}{minimal number of data points per concentration level}
}
\description{
Clean data
}
