% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D_ALB}
\alias{D_ALB}
\title{Targeted Proteomics data set Albumin}
\format{
\subsection{\code{D_ALB}}{

A data frame with 21 rows and 2 columns:
\describe{
\item{Concentration}{Concentration in fmol/µl}
\item{Measurement}{Measurement as ratio between response of stable isotope standard and native peptide}
}
}
}
\source{
Y. Mohammed, A. J. Percy, A. G. Chambers, C. H. Borchers, J. Proteome Res. 2015, 14, 1137.
}
\usage{
D_ALB
}
\description{
Measurement of the y8 fragment ion of the  peptide LVNEVTEFAK of Albumin (Mohammed et al., 2015).
TODO
}
\keyword{datasets}
