% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D_MFAP4}
\alias{D_MFAP4}
\title{Targeted Proteomics data set MFAP4}
\format{
\subsection{\code{D_MFAP4}}{

A data frame with 55 rows and 2 columns:
\describe{
\item{Concentration}{Concentration}
\item{Measurement}{Measurement}
}
}
}
\source{
TODO
}
\usage{
D_MFAP4
}
\description{
Details TODO
}
\keyword{datasets}
