% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_response_factors.R
\name{.calcResponseFactors}
\alias{.calcResponseFactors}
\title{Calculate Response factors}
\usage{
.calcResponseFactors(x, intercept, expConc)
}
\arguments{
\item{x}{\emph{data.frame}* \cr Data.frame containing data for a specific concentration level.}

\item{intercept}{\strong{numeric(1)} \cr Intercept of the linear model.}

\item{expConc}{\strong{numeric(1)} \cr  Expected concentration (known concentration value).}
}
\value{
vector of response factors for this specific concentration level
}
\description{
Function, which calculates a response factor for a single data point
}
\details{
Formula obtained from:  Green, J. M., A practical guide to analytical method validation.
}
