% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_plotting.R
\name{plotResponseFactors}
\alias{plotResponseFactors}
\title{Plot response factors}
\usage{
plotResponseFactors(RES, ylab = "Response Factor", xlab = "Concentration")
}
\arguments{
\item{RES}{\strong{list} \cr Results of \code{\link{CalibraCurve}}.}

\item{ylab}{\strong{character(1)} \cr y-axis label.}

\item{xlab}{\strong{character(1)} \cr x-axis label.}
}
\value{
A ggplot2 object containing the response factor plot.
}
\description{
Plot response factors
}
\examples{

data(RES_ALB)
plotResponseFactors(RES_ALB)
}
