% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_preprocessing.R
\name{readData}
\alias{readData}
\title{Read in data in different input formats}
\usage{
readData(
  data_path,
  filetype,
  conc_col,
  meas_col,
  sep = ",",
  dec = ";",
  header = TRUE,
  na.strings = c("NA", "NaN", "Filtered", "#NV"),
  sheet = 1
)
}
\arguments{
\item{data_path}{\strong{character(1)} \cr Path to the data file (.csv, .txt or .xlsx file).}

\item{filetype}{\strong{character(1)} \cr Type of input file: "csv" or "txt" or "xlsx".}

\item{conc_col}{\strong{integer(1)} \cr Column number of the concentration values.}

\item{sep}{\strong{character(1)} \cr The field separator, e.g. " " for blanks, "," for comma or "\t" for tab.}

\item{dec}{\strong{character(1)} \cr Decimal separator, e.g. "," for comma or "." for dot.}

\item{header}{\strong{logical(1)} \cr If TRUE, first line is counted as column names.}

\item{na.strings}{\strong{character} \cr Character vector of strings which are to be interpreted as NA.}

\item{sheet}{\strong{integer(1)} \cr Sheet number (only needed for xlsx files, default is to use the first sheet).}

\item{meas_cols}{\strong{integer} \cr Column number of the concentration values.}
}
\value{
data.frame with two numeric columns: Concentration and Measurement
}
\description{
Read in data in different input formats
}
